Implemented basic framework
Made a simple framework that consists of an abstract Master class, which is responsible for receiving packets from the simulator/SSL-vision and updating robot information. The ClientMaster derives from Master and handles communication specifically back to the simulator. Also made a Robot class that gives access to the robot's position and angle, while also containing methods for setting the target velocities.