Added command system
Added a command system for implementing robot "skills." Commands allow you to attach behavior to a robot and execute it synchronously with the run loop. Default commands can also be assingned to robots, which run when no other commands are active (this could be useful for a goalie bot, for instance).