Started on robot passing skill
Started developing the robot passing skill. Right now, the passing robot can get into a position where its kicker is aimed at the ball and the ball is in between the passing and receiving robot. The passing robot also avoids contact with the ball while getting into position, and if the position of the receiving robot changes, the passing robot makes an orbital motion around the ball to adjust.