Finished basic passing commands
Basic passing is now functional. I added kicker functionality on the robot, which is the last of the basic  robot features that needed to be implemented. I added multiple commands to enable the robot to align with the ball, approach it, shoot, and repeat. Right now, since this is set as the default command, robot 0 constantly shoots at robot 2 and chases after the ball, making for a fun dodgeball minigame :) also fixed a few bugs with CommandSeries and added an Init virtual function that's called when a Comamnd is assigned to a robot.