Implemented catching the ball
Implemented ball catching, and made a few improvements to the Command system and update loop. There were a few occasions where the simulation would update twice before a packet was sent, so the speed readings would be way off. So now the update loop tracks frames instead of time, and multiplies a constant delta time by the frames passed.